+++ main/moonraker_probe.c
@@ -8,6 +8,7 @@
 #include "lwip/sockets.h"
 
 #include <errno.h>
+#include <fcntl.h>
 #include <stdio.h>
 #include <string.h>
 #include <unistd.h>